% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-reset.R
\name{incr_template_number}
\alias{incr_template_number}
\title{Increase Template Number}
\usage{
incr_template_number()
}
\value{
A string of the new template number.
}
\description{
Increase Template Number
}
