% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_datas}
\alias{load_datas}
\title{Load Data}
\usage{
load_datas(sub = get_sub(), main = get_main(), env = calling_env())
}
\arguments{
\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{env}{The environment to load the objects into}
}
\description{
Load Data
}
