% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{load_object}
\alias{load_object}
\title{Load Object}
\usage{
load_object(x, type = "", main = get_main(), sub = get_sub(),
  env = calling_env())
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{type}{A string of the type of x.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\description{
Load Object
}

