% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_object_recursive}
\alias{load_object_recursive}
\title{Load Object Recursive}
\usage{
load_object_recursive(x, sub = get_sub(), top = TRUE,
  main = get_main())
}
\arguments{
\item{x}{The object to save.}

\item{sub}{A string of the path to the directory to save the object.}

\item{top}{A flag indicating whether to include objects in the top folder.}

\item{main}{A string of the main subfolder.}
}
\description{
Load Object Recursive
}
