% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_tables}
\alias{load_tables}
\title{Load tables}
\usage{
load_tables(x, main = get_main(), sub = "")
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\description{
List of all tables in sub and its subdirectories.
}
