% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_template_recursive}
\alias{load_template_recursive}
\title{Load Templates}
\usage{
load_template_recursive(x, sub = get_sub(), main = get_main())
}
\arguments{
\item{x}{The object to save.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}
}
\description{
Load Templates
}
