% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{md_tables}
\alias{md_tables}
\title{Markdown Templates}
\usage{
md_tables(headings = list(character(0)), drop = list(character(0)),
  report = get_report(), main = get_main(), sub = "", headers = c(3L,
  3L), locale = "en", ask = getOption("subfoldr.ask", TRUE))
}
\arguments{
\item{headings}{A list of named character vectors.}

\item{drop}{A list of character vectors specify the subfolders to drop.}

\item{report}{A string of the path to the report folder.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the subfolders to save the object (by default = "").}

\item{headers}{An integer vector of length 2 specify the header level for the first subfolders and the numer}

\item{locale}{A string of the locale.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}
}
\value{
A string of the report templates in markdown format ready for inclusion in a report.
}
\description{
Returns a string of templates in markdown format ready for inclusion in a report.
}
\details{
The names in the character vectors in headings indicate the new headings for each subfolder.
By default missing subfolders receive their current name with the first letter of each word capitalized.
Subfolders with the name "" do not receive a heading.
The order of the subfolders indicates the order in which they should appear.
By default missing subfolders appear in alphabetical order.
The first named character vector is applied to the highest level of subfolders starting at sub and so on.
The number of character vectors indicates the number of levels that should receive headings.
By default the highest level of subfolders are considered to be third order headings.

The elements in the character vectors in drop indicate the subfolders to exclude from the report.
Again the number of the character vector indicates the level to which it applies.
}
