% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object_subdirs}
\alias{object_subdirs}
\title{List Object Sub Directories}
\usage{
object_subdirs(main = get_main(), sub = get_sub())
}
\arguments{
\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\value{
A character vector of the names of the directories in the sub folder.
}
\description{
List Object Sub Directories
}
