% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{open_db}
\alias{open_db}
\title{Opens an connection to an sqlite3 database}
\usage{
open_db(x, new = NA, foreign_keys = TRUE, sub = get_sub(),
  main = get_main(), ask = getOption("subfoldr.ask", TRUE), ...)
}
\arguments{
\item{x}{A string of the name of the db to create (without the file extension).}

\item{new}{A flag indicating whether to enforce connection to an existing database (FALSE) or connection to a new database (TRUE).}

\item{foreign_keys}{A flag indicating whether to switch foreign keys on.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}

\item{...}{Additional arguments passed to pdf}
}
\description{
Opens an connection to an sqlite3 database
}
