% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-window.R
\name{open_window}
\alias{open_window}
\title{Open a new graphics window.}
\usage{
open_window(width = 6, height = width)
}
\arguments{
\item{width}{A number indicating the width in inches.}

\item{height}{A number indicating the height in inches.}
}
\description{
Open a new graphics window.
}

