% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{plot_exists}
\alias{plot_exists}
\title{Plot exists}
\usage{
plot_exists(x, sub = get_sub(), main = get_main())
}
\arguments{
\item{x}{The object to save.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}
}
\description{
Test whether a plot named x exists.
}
