% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-reset.R
\name{reset_template_number}
\alias{reset_template_number}
\title{Reset Template Number}
\usage{
reset_template_number()
}
\value{
A string of the new template number.
}
\description{
Reset Template Number
}
