% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{rm_datas}
\alias{rm_datas}
\title{Remove Data}
\usage{
rm_datas(sub = get_sub(), main = get_main(),
  ask = getOption("subfoldr.ask", TRUE))
}
\arguments{
\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}
}
\description{
Remove Data
}
