% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{rm_objects}
\alias{rm_objects}
\title{Remove Objects}
\usage{
rm_objects(sub = get_sub(), main = get_main(), recursive = TRUE,
  ask = getOption("subfoldr.ask", TRUE))
}
\arguments{
\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{recursive}{A flag indicating whether to recursively delete items.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}
}
\description{
Remove Objects
}
