% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_datas}
\alias{save_datas}
\title{Save data frames in the environ}
\usage{
save_datas(sub = get_sub(), main = get_main(), env = calling_env(),
  ask = getOption("subfoldr.ask", TRUE))
}
\arguments{
\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{env}{The environment to get the data from.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}
}
\value{
TRUE or FALSE
}
\description{
Save data frames in the environ
}
