% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_multiplot}
\alias{save_multiplot}
\title{Save Object with multiple viewports as .png
For each plot provided in 'plot' argument, there should be a corresponding viewport in 'vp' argument (where first plot object corresponds to first viewport object, etc.).}
\usage{
save_multiplot(x, sub = get_sub(), main = get_main(), caption = "",
  report = TRUE, width = NA_real_, height = NA_real_, dpi = 300,
  ask = getOption("subfoldr.ask", TRUE), plot, vp)
}
\arguments{
\item{x}{A string of the plot name.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{caption}{A string of the figure caption.}

\item{report}{A flag indicating to include the plot in reports.}

\item{width}{A number indicating the width of the plot in inches.}

\item{height}{A number indicating the height of the plot in inches.}

\item{dpi}{A number indicating the resolution of the .png.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}

\item{plot}{A list containing plot objects.}

\item{vp}{A list containing viewport objects.}
}
\description{
Save Object with multiple viewports as .png
For each plot provided in 'plot' argument, there should be a corresponding viewport in 'vp' argument (where first plot object corresponds to first viewport object, etc.).
}
