% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_plot}
\alias{save_plot}
\title{Save Object as .png}
\usage{
save_plot(x, sub = get_sub(), main = get_main(), caption = "",
  report = TRUE, width = NA_real_, height = NA_real_, dpi = 300,
  csv = TRUE, ask = getOption("subfoldr.ask", TRUE),
  plot = ggplot2::last_plot())
}
\arguments{
\item{x}{A string of the plot name.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{caption}{A string of the figure caption.}

\item{report}{A flag indicating to include the plot in reports.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height}{A number indicating the height of the plot in inches.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{csv}{A flag indicating whether to save a csv of the plot data.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}

\item{plot}{Plot to save, defaults to last plot displayed.}
}
\description{
Save Object as .png
}
