% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{save_table}
\alias{save_table}
\title{Save Object as .csv}
\usage{
save_table(x, type = get_type(), main = get_main(), sub = get_sub(),
  x_name = NULL, caption = "", ask = getOption("subfoldr.ask", TRUE))
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{type}{A string of the type of x.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}

\item{x_name}{An optional string of the name to use.}

\item{caption}{A string of the figure caption.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}
}
\value{
The object x or TRUE or FALSE is x is missing.
}
\description{
Save Object as .csv
}

