% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-reset.R
\name{set_plot_number}
\alias{set_plot_number}
\title{Set Plot Number}
\usage{
set_plot_number(x)
}
\arguments{
\item{x}{A count.}
}
\value{
A string of the new plot number.
}
\description{
Set Plot Number
}
