% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-reset.R
\name{set_table_number}
\alias{set_table_number}
\title{Set Table Number}
\usage{
set_table_number(x)
}
\arguments{
\item{x}{A count.}
}
\value{
A string of the new table number.
}
\description{
Set Table Number
}
