% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-reset.R
\name{set_template_number}
\alias{set_template_number}
\title{Set Template Number}
\usage{
set_template_number(x)
}
\arguments{
\item{x}{A count.}
}
\value{
A string of the new template number.
}
\description{
Set Template Number
}
