% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdirs.R
\name{subdirs_tables}
\alias{subdirs_tables}
\title{List Table Sub Directories}
\usage{
subdirs_tables(main = get_main(), sub = get_sub())
}
\arguments{
\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\value{
A character vector of the names of the directories in the sub folder.
}
\description{
List Table Sub Directories
}
