% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{table_exists}
\alias{table_exists}
\title{Table exists}
\usage{
table_exists(x, main = get_main(), sub = get_sub())
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\description{
Test whether a table named x exists.
}
