% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_subdirs}
\alias{table_subdirs}
\title{List Table Sub Directories}
\usage{
table_subdirs(main = get_main(), sub = get_sub())
}
\arguments{
\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\value{
A character vector of the names of the directories in the sub folder.
}
\description{
List Table Sub Directories
}

