% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{template_exists}
\alias{template_exists}
\title{Template exists}
\usage{
template_exists(x, sub = get_sub(), main = get_main())
}
\arguments{
\item{x}{The object to save.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}
}
\description{
Test whether an template named x exists.
}
