% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_subdirs}
\alias{template_subdirs}
\title{List Template Sub Directories}
\usage{
template_subdirs(main = get_main(), sub = get_sub())
}
\arguments{
\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\value{
A character vector of the names of the directories in the sub folder.
}
\description{
List Template Sub Directories
}

