% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{load_object}
\alias{load_object}
\title{Load Object}
\usage{
load_object(x_name = NULL, type = "", main = get_main(),
  sub = get_sub())
}
\arguments{
\item{x_name}{An optional string of the name to use.}

\item{type}{A string of the type of x.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\description{
Load Object
}

