% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{load_object}
\alias{load_object}
\title{Load Object}
\usage{
load_object(x, main = get_main(), sub = get_sub(), is = is.data.frame,
  env = calling_env())
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}

\item{is}{A function returning TRUE or FALSE to indicate whether to save particular objects.
Ignored unless x is missing.}

\item{env}{The environment to load the objects into if x is missing.}
}
\description{
Load Object
}

