% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_objects}
\alias{load_objects}
\title{Load objects}
\usage{
load_objects(x, main = get_main(), sub = get_sub(), data = TRUE)
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}

\item{data}{A flag indicating whether to only load objects inheriting from data.frame.}
}
\description{
List of all data objects in sub and its subdirectories.
}
