% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/table.R
\name{load_table}
\alias{load_table}
\title{Load table}
\usage{
load_table(x = NULL, type = "", main = get_main(), sub = get_sub())

load_table(x = NULL, type = "", main = get_main(), sub = get_sub())
}
\description{
Load table

Load table
}

