% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_table}
\alias{load_table}
\title{Load table}
\usage{
load_table(x, main = get_main(), sub = get_sub(), env = calling_env())
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}

\item{env}{The environment to load the objects into if x is missing.}
}
\description{
Load table
}
