% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_data}
\alias{save_data}
\title{Save Data}
\usage{
save_data(x, sub = get_sub(), main = get_main(), x_name = NULL,
  ask = getOption("subfoldr.ask", TRUE))
}
\arguments{
\item{x}{The data frame to save. If missing saves all data frames in calling env.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{x_name}{An optional string of the name to use.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}
}
\value{
The object x or TRUE or FALSE is x is missing.
}
\description{
Save Data
}
