% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save Object as .png}
\usage{
save_plot(x, type = get_type(), main = get_main(), sub = get_sub(),
  width = NA, height = NA, dpi = 300, caption = "",
  ask = getOption("subfoldr.ask", TRUE), plot = ggplot2::last_plot())
}
\arguments{
\item{x}{A string of the plot name.}

\item{type}{A string of the type of x.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height}{A number indicating the height of the plot in inches.}

\item{dpi}{Plot resolution. Applies only to raster output types.}

\item{caption}{A string of the figure caption.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}

\item{plot}{Plot to save, defaults to last plot displayed.}
}
\description{
Save Object as .png
}

