% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_table}
\alias{save_table}
\title{Save Object as .csv}
\usage{
save_table(x, x_name = NULL, sub = get_sub(), main = get_main(),
  caption = "", report = TRUE, ask = getOption("subfoldr.ask", TRUE))
}
\arguments{
\item{x}{The object to save.}

\item{x_name}{An optional string of the name to use.}

\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}

\item{caption}{A string of the figure caption.}

\item{report}{A flag indicating to include the plot in reports.}

\item{ask}{A string indicating whether to ask before creating a sub directory.}
}
\value{
The object x.
}
\description{
Save Object as .csv
}
