% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdirs.R
\name{subdirs_tables}
\alias{subdirs_tables}
\title{List Table Sub Directories}
\usage{
subdirs_tables(sub = get_sub(), main = get_main())
}
\arguments{
\item{sub}{A string of the path to the directory to save the object.}

\item{main}{A string of the main subfolder.}
}
\value{
A character vector of the names of the directories in the sub folder.
}
\description{
List Table Sub Directories
}
