% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{template_exists}
\alias{template_exists}
\title{Template exists}
\usage{
template_exists(x, main = get_main(), sub = get_sub())
}
\arguments{
\item{x}{The object to save. If missing saves all objects in calling env.}

\item{main}{A string of the main subfolder.}

\item{sub}{A string of the path to the directory to save the object.}
}
\description{
Test whether an template named x exists.
}
